feat: combo toolbar
