+++ src/components/layout-panel/bottom-bar/action-buttons/styles/action-buttons.module.css
@@ -1,5 +1,6 @@
 .button {
     display: flex;
+    flex: 0 0 auto;
     align-items: center;
     gap: 4px;
     min-block-size: 28px;
@@ -13,6 +14,7 @@
     color: var(--colors-grey700);
     background-color: var(--colors-grey-050);
     box-shadow: none;
+    white-space: nowrap;
 }
 
 .button:disabled,
@@ -62,6 +64,8 @@
 
 .splitButton {
     display: flex;
+    flex: 0 0 auto;
+    flex-wrap: nowrap;
 }
 
 .button.splitLeft {
