--- src/components/layout-panel/bottom-bar/styles/bottom-bar.module.css
@@ -1,28 +1,0 @@
-.bottomBar {
-    padding: 6px;
-    background-color: var(--colors-white);
-    border-block-start: 1px solid var(--colors-grey300);
-    box-sizing: border-box;
-}
-
-.bottomBar.empty,
-.bottomBar.loading {
-    block-size: 29px;
-}
-
-.container {
-    display: flex;
-    align-items: center;
-    justify-content: flex-start;
-    gap: var(--spacers-dp12);
-    animation: fadeIn 0.4s ease-out forwards;
-}
-
-@keyframes fadeIn {
-    0% {
-        opacity: 0;
-    }
-    100% {
-        opacity: 1;
-    }
-}
