+++ src/components/layout-panel/layout-panel.tsx
@@ -2,7 +2,6 @@ import { useAppSelector } from '@hooks'
 import { getUiLayoutPanelVisible } from '@store/ui-slice'
 import { type FC } from 'react'
 import { Axes } from './axes'
-import { BottomBar } from './bottom-bar/bottom-bar'
 import classes from './styles/layout-panel.module.css'
 import { TopBar } from './top-bar/top-bar'
 
@@ -13,7 +12,6 @@ export const LayoutPanel: FC = () => {
         <div className={classes.panel}>
             <TopBar />
             <Axes />
-            <BottomBar />
         </div>
     ) : null
 }
