+++ src/components/layout-panel/top-bar/top-bar.tsx
@@ -1,6 +1,11 @@
+import { CustomValueButton } from '@components/layout-panel/bottom-bar/action-buttons/custom-value-button'
+import { EnrollmentButton } from '@components/layout-panel/bottom-bar/action-buttons/enrollment-button'
+import { EventButton } from '@components/layout-panel/bottom-bar/action-buttons/event-button'
+import { TrackedEntityInstanceButton } from '@components/layout-panel/bottom-bar/action-buttons/tracked-entity-instance-button'
 import { useAppSelector } from '@hooks'
 import { getDataSourceId } from '@store/dimensions-selection-slice'
 import { getIsVisualizationLoading } from '@store/loader-slice'
+import { getVisUiConfigVisualizationType } from '@store/vis-ui-config-slice'
 import cx from 'classnames'
 import { type FC } from 'react'
 import { ExpandedLayoutPanelToggler } from './expanded-layout-panel-toggler'
@@ -11,6 +16,7 @@ import { VisualizationTypeSelector } from './visualization-type-selector/visuali
 export const TopBar: FC = () => {
     const dataSourceId = useAppSelector(getDataSourceId)
     const isVisualizationLoading = useAppSelector(getIsVisualizationLoading)
+    const visualizationType = useAppSelector(getVisUiConfigVisualizationType)
 
     return (
         <div
@@ -20,8 +26,24 @@ export const TopBar: FC = () => {
         >
             {!isVisualizationLoading && (
                 <div className={classes.container}>
-                    <VisualizationTypeSelector />
-                    <OptionsButton />
+                    {dataSourceId && (
+                        <div
+                            className={classes.updateButtons}
+                            data-test="update-buttons"
+                        >
+                            <EventButton />
+                            <EnrollmentButton />
+                            {visualizationType === 'PIVOT_TABLE' ? (
+                                <CustomValueButton />
+                            ) : (
+                                <TrackedEntityInstanceButton />
+                            )}
+                        </div>
+                    )}
+                    <div className={classes.visualizationControls}>
+                        <VisualizationTypeSelector />
+                        <OptionsButton />
+                    </div>
                     <div className={classes.containerSpacer} />
                     {dataSourceId && <ExpandedLayoutPanelToggler />}
                 </div>
