+++ src/components/layout-panel/top-bar/visualization-type-selector/styles/visualization-type-selector.module.css
@@ -5,6 +5,7 @@
 
 .button {
     display: flex;
+    flex: 0 0 auto;
     align-items: center;
     gap: var(--spacers-dp4);
     padding: 6px 8px;
@@ -13,6 +14,7 @@
     border-radius: 5px;
     background-color: var(--colors-white);
     cursor: pointer;
+    white-space: nowrap;
 }
 
 .button:hover,
